- mypy substation runs again wherever numpy 2.5 or later is installed: its stubs use Python 3.12 syntax, and python_version = "3.10" stopped mypy before it checked any code, as on any fresh install on Python 3.12 or later - mypy now checks for the Python it runs under; running it under 3.10 checks the oldest supported Python, and numpy 2.5 does not install there anyway - Fixed the errors that showed up: the noise floor estimate returned numpy floats where float is declared (seen on 3.10), and the soft limiter's output became float64 (seen with numpy 2.5) - Checked clean on Python 3.12 with numpy 2.4 and 2.5, and on 3.10 with numpy 2.2, where the test suite also passes - Dropped the comment calling the setup strict, since no strict options are set

+++ substation/recording.py
@@ -839,7 +839,7 @@ class ChannelRecorder:
 
 		# Apply soft limiter using precomputed parameters
 		if samples.size > 0 and self.soft_limit_drive > 0:
-			samples = numpy.tanh(samples * self.soft_limit_drive) * self.soft_limit_scale
+			samples = (numpy.tanh(samples * self.soft_limit_drive) * self.soft_limit_scale).astype(numpy.float32, copy=False)
 
 		# It will automatically convert to int16 internally
 		with self._write_lock:
